Make options document stable
Some options' default value includes hash code or timestamp. As a result, the document can be changed even when the API is not changed. This commit sets contant values of `defaultText` for these options so that they will not depend on hash code or timestamp any more and will become stable.